+++ connect/startup.sh
@@ -40,11 +40,14 @@ elif test -f "$RSC_LICENSE_FILE_PATH"; then
     rm -f /var/lib/rstudio-connect/*.lic
     cp "${RSC_LICENSE_FILE_PATH}" /var/lib/rstudio-connect/license.lic
     chmod g-rwx,g-rwx /var/lib/rstudio-connect/license.lic
+elif ls /var/lib/rstudio-connect/*.lic >/dev/null 2>&1; then
+    echo "Detected a license file in /var/lib/rstudio-connect/*.lic."
 fi
 
 # ensure these cannot be inherited by child processes
 unset RSC_LICENSE
 unset RSC_LICENSE_SERVER
+unset RSC_LICENSE_FILE_PATH
 
 # Start RStudio Connect
 /opt/rstudio-connect/bin/connect --config /etc/rstudio-connect/rstudio-connect.gcfg